requirements.txt
absl-py==0.13.0
addict==2.4.0
anykeystore==0.2
apex==0.1
appdirs==1.4.4
apptools==5.1.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
black==21.7b0
bleach==3.3.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.1
click==8.0.1
configobj==5.0.6
cryptacular==1.5.5
cycler==0.10.0
Cython==0.29.23
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
descartes==1.1.0
#-e git+https://github.com/Vegeta2020/CIA-SSD.git@7a45493651478ef4915e9d13861d74ff6317cae0#egg=det3d
easydict==1.9
entrypoints==0.3
fire==0.4.0
flake8==3.9.2
future==0.18.2
google-auth==1.32.1
google-auth-oauthlib==0.4.4
GPUtil==1.4.0
graphviz==0.17
greenlet==1.1.0
grpcio==1.38.1
hupper==1.10.3
idna==3.2
#-e git+https://github.com/jackd/ifp-sample.git@d506204da70f2b33c4301b2378b0d623c2a4802a#egg=ifp_sample
imagecorruptions==1.1.2
imageio==2.9.0
importlib-metadata==4.6.1
importlib-resources==5.2.0
iniconfig==1.1.1
iou3d==0.0.0
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
jetson-stats==3.1.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
kornia==0.5.7
llvmlite==0.33.0
lyft-dataset-sdk==0.0.8
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mayplotlib==0.0.1
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.4.0
numba==0.50.1
numpy==1.19.3
nuscenes-devkit==1.1.5
oauthlib==3.1.1
objgraph==3.5.0
open3d-python==0.3.0.0
opencv-contrib-python==4.5.3.56
opencv-python==4.3.0.38
packaging==21.0
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
PasteDeploy==2.1.1
pathspec==0.9.0
pbkdf2==1.3
#-e git+https://github.com/open-mmlab/OpenPCDet.git@26a16123cf820e29ed565f4ad53964c66f3ea030#egg=pcdet
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
plaster==1.0
plaster-pastedeploy==0.7
plotly==5.1.0
pluggy==0.13.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.7.0
pycocotools==2.0.2
pycodestyle==2.7.0
pycparser==2.20
pyface==7.3.0
pyflakes==2.3.1
Pygments==2.9.0
pyparsing==2.4.7
pyquaternion==0.9.9
pyramid==2.0
pyramid-mailer==0.15.1
pyrsistent==0.18.0
pytest==6.2.4
pytest-runner==5.3.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
regex==2021.7.6
repoze.sendmail==4.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.17.2
scikit-learn==0.22.1
scipy==1.5.4
Send2Trash==1.7.1
Shapely==1.6.4.post2
six==1.16.0
#spconv @ file:///home/mj/YOLO/source/ssd/spconv/dist/spconv-1.0-cp36-cp36m-linux_aarch64.whl
SQLAlchemy==1.4.22
tenacity==8.0.1
tensorboard==2.2.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
termcolor==1.1.0
terminado==0.10.1
terminaltables==3.1.0
testpath==0.5.0
tifffile==2020.9.3
toml==0.10.2
tomli==1.2.0
#torch @ file:///home/mj/YOLO/torch-1.6.0-cp36-cp36m-linux_aarch64.whl
torchsummary==1.5.1
torchvision==0.7.0
tornado==6.1
tqdm==4.61.2
traitlets==4.3.3
traits==6.2.0
traitsui==7.2.1
transaction==3.0.1
translationstring==1.4
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.6
velruse==1.1.1
venusian==3.0.0
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.7
Werkzeug==2.0.1
widgetsnbextension==3.5.1
WTForms==2.3.3
wtforms-recaptcha==0.3.2
zipp==3.5.0
zope.deprecation==4.4.0
zope.interface==5.4.0
zope.sqlalchemy==1.5
